# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LATTICE v1.0",
    "description": "Verified financial orchestration. All execution is SANDBOX.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/funding-sources": {
      "get": {
        "summary": "Funds",
        "operationId": "funds_api_v1_funding_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Fund",
        "operationId": "add_fund_api_v1_funding_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundingInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/funding-sources/{id}": {
      "patch": {
        "summary": "Patch Fund",
        "operationId": "patch_fund_api_v1_funding_sources__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundingPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/obligations": {
      "get": {
        "summary": "Obligations",
        "operationId": "obligations_api_v1_obligations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Obligation",
        "operationId": "add_obligation_api_v1_obligations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObligationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/obligations/{id}": {
      "get": {
        "summary": "Obligation",
        "operationId": "obligation_api_v1_obligations__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Obligation",
        "operationId": "patch_obligation_api_v1_obligations__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObligationPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/obligations/{id}/verify-beneficiary": {
      "post": {
        "summary": "Verify Beneficiary",
        "operationId": "verify_beneficiary_api_v1_obligations__id__verify_beneficiary_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeneficiaryVerification"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfer-routes": {
      "get": {
        "summary": "Routes",
        "operationId": "routes_api_v1_transfer_routes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add Route",
        "operationId": "add_route_api_v1_transfer_routes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permissions": {
      "get": {
        "summary": "Permissions",
        "operationId": "permissions_api_v1_permissions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Grant",
        "operationId": "grant_api_v1_permissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/permissions/{id}": {
      "delete": {
        "summary": "Revoke",
        "operationId": "revoke_api_v1_permissions__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeneficiaryVerification": {
        "properties": {
          "beneficiary": {
            "type": "string",
            "pattern": "^[A-Z0-9_-]{3,80}$",
            "title": "Beneficiary"
          },
          "repeat_beneficiary": {
            "type": "string",
            "title": "Repeat Beneficiary"
          },
          "confirmation_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 20,
            "title": "Confirmation Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "beneficiary",
          "repeat_beneficiary",
          "confirmation_note"
        ],
        "title": "BeneficiaryVerification"
      },
...
      "FundingInput": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Label"
          },
          "source_type": {
            "type": "string",
            "enum": [
              "STUDENT_BALANCE",
              "PARENT_SUPPORT",
              "SCHOLARSHIP",
              "EMERGENCY_RESERVE",
              "OTHER"
            ],
            "title": "Source Type"
          },
          "amount": {
            "type": "number",
            "maximum": 1000000000.0,
            "minimum": 0.0,
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD",
              "CNY",
              "VND",
              "GBP",
              "JPY",
              "KRW",
              "SGD",
              "HKD",
              "AUD",
              "CAD",
              "CHF",
              "THB"
            ],
            "title": "Currency"
          },
          "available_from": {
            "type": "string",
            "format": "date",
            "title": "Available From"
          },
          "availability_status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "PLANNED",
              "EXPECTED",
              "CONDITIONAL",
              "LOCKED",
              "DISPUTED"
            ],
            "title": "Availability Status"
          },
          "restriction_type": {
            "type": "string",
            "enum": [
              "UNRESTRICTED",
              "EMERGENCY",
              "TUITION_ONLY"
            ],
            "title": "Restriction Type",
            "default": "UNRESTRICTED"
          },
          "minimum_remaining_balance": {
            "type": "number",
            "minimum": 0.0,
            "title": "Minimum Remaining Balance",
            "default": 0
          },
          "confirmation_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Confirmation Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "source_type",
          "amount",
          "currency",
          "available_from",
          "availability_status",
          "confirmation_note"
        ],
        "title": "FundingInput"
      },
      "FundingPatch": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "available_from": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available From"
          },
          "availability_status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "AVAILABLE",
                  "PLANNED",
                  "EXPECTED",
                  "CONDITIONAL",
                  "LOCKED",
                  "DISPUTED"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Availability Status"
          },
          "confirmation_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Confirmation Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirmation_note"
        ],
        "title": "FundingPatch"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObligationInput": {
        "properties": {
          "label": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Label"
          },
          "type": {
            "type": "string",
            "enum": [
              "TUITION",
              "HOUSING_DEPOSIT",
              "RENT",
              "INSURANCE",
              "OTHER"
            ],
            "title": "Type"
          },
          "beneficiary": {
            "type": "string",
            "pattern": "^[A-Z0-9_-]{3,80}$",
            "title": "Beneficiary"
          },
          "amount": {
            "type": "number",
            "maximum": 1000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD",
              "CNY",
              "VND",
              "GBP",
              "JPY",
              "KRW",
              "SGD",
              "HKD",
              "AUD",
              "CAD",
              "CHF",
              "THB"
            ],
            "title": "Currency"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "priority": {
            "type": "string",
            "enum": [
              "CRITICAL",
              "HIGH",
              "NORMAL",
              "OPTIONAL"
            ],
            "title": "Priority"
          },
          "partial_payment_allowed": {
            "type": "boolean",
            "title": "Partial Payment Allowed",
            "default": false
          },
          "minimum_payment": {
            "type": "number",
            "minimum": 0.0,
            "title": "Minimum Payment",
            "default": 0
          },
          "confirmation_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Confirmation Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "label",
          "type",
          "beneficiary",
          "amount",
          "currency",
          "due_date",
          "priority",
          "confirmation_note"
        ],
        "title": "ObligationInput"
      },
      "ObligationPatch": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000000000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "confirmation_note": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 8,
            "title": "Confirmation Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "confirmation_note"
        ],
        "title": "ObligationPatch"
      },
      "PermissionInput": {
        "properties": {
          "target_actor_id": {
            "type": "string",
            "title": "Target Actor Id"
          },
          "resource_type": {
            "type": "string",
            "enum": [
              "FUNDING",
              "OBLIGATION",
              "PLAN",
              "ACTION"
            ],
            "title": "Resource Type"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "permission_type": {
            "type": "string",
            "enum": [
              "VIEW_SHARED_OBLIGATION",
              "VIEW_OWN_CONTRIBUTION",
              "APPROVE_OWN_FUNDS",
              "VIEW_PRIVATE_FINANCES",
              "APPROVE_PLAN",
              "APPROVE_ACTION"
            ],
            "title": "Permission Type"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "target_actor_id",
          "resource_type",
          "resource_id",
          "permission_type"
        ],
        "title": "PermissionInput"
      },
      "RouteInput": {
        "properties": {
          "provider_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Provider Name"
          },
          "from_currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD",
              "CNY",
              "VND",
              "GBP",
              "JPY",
              "KRW",
              "SGD",
              "HKD",
              "AUD",
              "CAD",
              "CHF",
              "THB"
            ],
            "title": "From Currency"
          },
          "to_currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD",
              "CNY",
              "VND",
              "GBP",
              "JPY",
              "KRW",
              "SGD",
              "HKD",
              "AUD",
              "CAD",
              "CHF",
              "THB"
            ],
            "title": "To Currency"
          },
          "fixed_fee": {
            "type": "number",
            "minimum": 0.0,
            "title": "Fixed Fee"
          },
          "percentage_fee": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Percentage Fee"
          },
          "fx_rate": {
            "type": "number",
            "maximum": 1000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Fx Rate"
          },
          "fx_markup": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Fx Markup"
          },
          "min_transfer": {
            "type": "number",
            "minimum": 0.0,
            "title": "Min Transfer"
          },
          "max_transfer": {
            "type": "number",
            "maximum": 1000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Transfer"
          },
          "settlement_p50_days": {
            "type": "integer",
            "maximum": 30.0,
            "minimum": 0.0,
            "title": "Settlement P50 Days"
          },
          "settlement_p95_days": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 0.0,
            "title": "Settlement P95 Days"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider_name",
          "from_currency",
          "to_currency",
          "fixed_fee",
          "percentage_fee",
          "fx_rate",
          "fx_markup",
          "min_transfer",
          "max_transfer",
          "settlement_p50_days",
          "settlement_p95_days"
        ],
        "title": "RouteInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}